Conflicts:
	app/aboot/aboot.c
	app/aboot/fastboot.c
	app/aboot/recovery.c
	include/platform.h
	include/stdio.h
	include/stdlib.h
	include/target.h
	kernel/timer.c
	lib/debug/debug.c
	lib/libfdt/fdt.c
	lib/libfdt/fdt_ro.c
	lib/libfdt/fdt_strerror.c
	lib/libfdt/fdt_sw.c
	lib/openssl/crypto/rules.mk
	platform/msm8226/acpuclock.c
	platform/msm8226/msm8226-clock.c
	platform/msm_shared/board.c
	platform/msm_shared/qgic.c
	platform/msm_shared/qpic_nand.c
	platform/msm_shared/rules.mk
	target/mdm9640/rules.mk

+++ platform/init.c
@@ -109,8 +109,9 @@ __WEAK uint32_t platform_get_sclk_count(void)
 	return 0;
 }
 
-__WEAK void clock_config_cdc(uint8_t slot)
+__WEAK void clock_config_cdc(uint32_t slot)
 {
+
 }
 
 __WEAK int get_target_boot_params(const char *cmdline, const char *part,
